The app will be an interactive form of the Survival Guidebook provided by Emergency Zone. https://www.emergencyzone.com/freepreparednessguidebook_a/275.htm
There will be a separate activity for each chapter, like the Miwock App.
You will learn the answers to  the quiz by reading the book.

If you answer incorrectly, it will give you a prompt to get the right answer. You can then choose to keep going, or try again.
Each question can have a link to where the answer is in the text (if Easy Mode is turned on).